Patched GPU error
The original code has a hardcoded parameter when invoking the docker images that tells it to look for the "second" GPU (device=1). This will break on systems with only 1 GPU (because the GPU is zero indexed). This sets --gpus=all, fixing the error.

This only applies to two files: pipeline_full_en.sh and pipeline_sample_en_m36.sh.